[Lib][Qt] Made the wallet's 'on_history' callback pass self.wallet
This drastically improves performance if multiple wallets are open, as
now each GUI window only responds to the on_history even for its own
wallet.

Note that fx and quotes are global on_history events and don't pass a
wallet, so the GUI always responds to those.

This is an optimization for CashShuffle that makes sense also for us in
regular.