CashShuffle: Fix #70; Shuffle tx description lost if last player is slow
It's possible for clients to time out in phase4 after sending their
signatures to the last player if the last player is very slow in signing
and responding to the other players.

The situation that arises is the last player finishes off the shuffle
and broadcasts the tx.

The other players believe the shuffle to have ended with a timeout.
However in reality the shuffle completes and the "Shuffle xx -> yy"
label is lost in the aether.

This commit is a workaround to that situation.

We "remember" the tentative shuffle tx and then monitor incoming tx's
for a tx that looks like what we were expecting.

This is all done with attention to not eating up too much CPU and only
examining incoming tx's when we timed out in phase4 and we believe it's
possible for the failed shuffle to have completed.