[Qt] Fixups to QR Window and QRDialog
- QRDialog was saving THE WHOLE SCREEN to png or clipboard. This has
been fixed
- QRWindow from the receive tab had funky logic.  I made it auto-kill
and remove itself whenever it is closed (no reason for it to stick around
in memory when closed). It gets re-created each time
- Fixed some unorthodox usage of the mouse cursor handling for the QR
code in the receive tab.  It kept pushing a new cursor onto the stack
each time which is out of spec and risks running out of mouse cursor
stack space.  Instead, we setCursor() on the QR widget.

... and other fixups related to this.