Fix #85 -- Closing the last wallet window now kills the pool windows
Before it was possible for the app to stay open if the tiny pool windows
(which might be obscured by other windows on the desktop) were still
open.

This was unexpected and strange behavior.

So now we trap the plugin on_close_window hook and if there are no more
wallet windows, kill all the pool windows.

Thanks for @emergent_reasons for this bug report.