Follow-up to last commit -- fix quirks of pool windows
It was possible to unload cashshuffle plugin and reload it and then the
pool windows woudld stay up on last wallet close. This has been
addressed by not using a queued connection on the close event.

Also it was possible to minimize a pool window and then "View pools..."
should unminimize it, but it didn't.  This has also been fixed.