Make "is_coin_shuffled" return "True" for ANY tx output amounts (#86)
... previously it was just the 'scale' amounts: 10.0, 1.0, 0.1, etc.

This is in preparation for supporting #68 -- "Use minimum amount for shuffle tx"

Now we improved the code to make any UTXO that appears to come from a shuffled transaction (based on transaction shape) be categorized as shuffled. The rules are as follows:

UTXO is shuffled if:

comes from a tx where there are at least 2 other outputs of the same amount
comes from a tx where if you group the outputs by amount, the utxo in question belongs to the longest (or tied for longest) amount-group. The length of this amount group must also be >= 3.
for the grouping by amount, all output addresses for this group must be different/unique.
At least 3 inputs to the tx the UTXO comes from
Note that we could make this check much more restrictive (such as checking that the inputs come from unique addresses -- but I don't think that's necessary and it would make the code much slower to do that since inputs in electron cash's wallet code lack the address, so you'd have to go look it up for each input).