[Qt] Layout/Alignment fixups to amounts in coins tab #1180
- Coins tab 'amounts' were misaligned. This was hard to read esp. for
wallets with lots of coins (such as CashShuffle wallets)

- While I was at it I noticed that frozen addresses look funky in the
coins tab in dark mode -- modified the color for dark scheme so it looks
ok.

- Slight performance fixups to Coins tab "on_update" refresher method to
used cached self.monospaceFont and self.colors rather than constructing
a QColor and QFont for *each item* in the list (this should help on
large wallets such as the eatbch mega-wallet).

Thanks Mengerian and markblundeberg !