[Qt] Fixups - Fixed Receive tab focus policy
In the "Receive" tab:

It was impossible to select the address and hit Ctrl-C (Cmd-C) to copy
the text if some other widget had focus.  This is unexpected, buggy
behavior.

This was due to the Address lineedit having "NoFocus" as its focus
policy.

There's no good reason for it to have no focus.  It's already readonly.

Users complained so I fixed.