Fix #83 - Add context menu to history tab to hide shuffle txs
Quick and dirty solution to #83 -- a simple toggle is added to the
right-click context menu to suppress (hide) all Shuffle tx's from the
history, app-wide.

Doing something more sophisticated is left up to a future release.  Also
-- implementing too much fancy logic in the history list significantly
affects performance on large wallets so this code is extra-critical to
get right.