Advanced config option: Raise the limit beyond 50 BCH (and also specify a lower limit) (#90)
* Update code to support scales up to 10,000 BCH

As well as getting the lower and upper bound on coin shuffling from the
config.

TODO: UI to configure this

* Added BackgroundShufflingThread.reset_lower_and_upper_bound_to_defaults

This will be used by the UI

* Added "Coin selection settings..." window

This window allows power users to specify low and upper amount cutoffs
for shuffling.  The defaults are 0.0001027 BCH and 50 BCH respectively.

Tiers up to 1000 BCH are also supported now.