Small performance tune-up for BackgrounShufflingThread
Refactored BackgroundShufflingThread coin selection to be more
efficient. It picks the set of coins once in the outer loop, applying
the configured LOWER_BOUND and UPPER_BOUND, then for each scale the set
is narrowed. This avoids lots of redundant calls and lookups within the
inner per-scale coin selector.  Won't really be noticed except on
wallets with very large numbers of UTXOs -- but it's worth doing as the
background process runs periodically to poll the state of the shuffles
and it filters coins each time it thinks there may be work to do.

I also made the "set defaults" button in the "Coin selectionsettings..."
window confirm it actually set defaults in the popup message. It also
clears the conf keys for min/max altogether (in case defaults change in
a future version) when you switch back to defaults.