Fully close wallet + Many PyQt memory leaks fixed (#1184)
On window close the wallet stays open and continues to synchronize 
needlessly. This is not expected UX. It also is a waste of resources 
and slows down the app. If you had several wallets open but then 
closed them the app would thus leak resources.

This has now been addressed in this commit. Closing a wallet shuts 
down all of its resources, including removing all of its callbacks for 
subscriptions and cancelling all pending network sends from that 
wallet (they are tracked down based on callback method supplied to
the request).

Since we can't "unsubscribe" from ElectrumX due to protocol 
limitations, we will get harmlessly notified on tx's in the network layer
for a closed wallet, but the notifications are now ignored.

When we switch servers, the subscription set sent to a new server 
will not include the closed wallet's addresses.

Also in this commit: I noticed when a window closes, the system tray 
QMenu items stick around as top level widgets (qApp.topLevelWidgets() 
in console reveals this). This has been addressed, plus some other 
leaked resource issues on the Qt side have been addressed.