[Qt] AddressList.update: resotre multiple selection after refresh
Previously if you selected more than 1 address at once the the widget
refreshed via 'update()' (badly named method as it clashes with
QtWidget.update.. but it is what it is), then your multiple selection
would be collapsed to 1 item (the most recently selected).

This has been addresses in this commit.