[Lib] Made format_satoshis run 50% faster with 1 simple trick
util.format_satoshis is called very often so it's a performance-critical
function. However, it was doing something slow -- namely looking up the
locale decimal point character each time it was called.  The locale doesn't
change during the app's lifetime (we don't support changing locales or
languages at runtime -- it requires an app restart) -- so doing it this
way involved needless overhead.

Now, we cache the decimal-point character lookup from localeconv() and use
that for each call, which turns out to be a 50% speedup for the
execution time of the format_satoshis function.

This speedup helps with very large wallets that have large history lists
and call this function often to generate the balance text string for the
UI.