[Lib] 1 last bug in wallet when deleting pk form imported wallet
The wallet and verifier weren't sane after deleting a key from a private
key imported wallet.  The verifier needs to be told of all the tx's that
were eviscerated so that if a pk is added again, it re-verifies.

(This avoids the UI bug where the UI thinks it's unverified and the
verifier never reverifies).

Fixed.