[Lib] Undo part of commit 65ef3ff5a585e3aa5ff318a85d191a0ee17aa124
Well, it turns out keeping the tx's we've seen in storage is actually a
good thing, on second thought. It helps with performance when switching
between forks so you don't re-download everything.

There is however a bug in undo_verifications() in that it doesn't check the
block header hash when deciding if a verificaton should be remove
(instead it checks timestamp!).

So, on chain reorg or fork switch, removing the tx's from the verified
list and the transaction history in storage was a bit heavy handed.

We will keep them around. The real fix will be for us to upgrade the
wallet storage format to also remember block hash when it stores a
verified tx hash.