Network: 'updated' callback split into: 'wallet_updated' , 'blockchain_updated' (#1187)
* Refactor: 'updated' callback -> 'wallet_updated' , 'blockchain_updated'

Also 'verified' is now 'verified2'.

Old callbacks are still supported but will cause warning spam to console using print_error.

This is a much needed change.  The 'updated' callback was overloaded to
indicate the wallet history was updated as well as the blockchain was
updated.

This was too large a granularity for what we need.

The 'verified' callback didn't indicate which wallet was verified (so *all* wallets needed to refresh, wasting valuable CPU cycles).

Since the Qt gui can have multiple wallet windows open, this led to a
situation where the GUI wallet windows were updating themselves multiple
times per wallet that got a new history and/or verification, even if the update didn't
concern their particular wallet.

So -- we split up the ambiguous 'updated' event into 'blockchain_updated'
and 'wallet_updated'.

blockchain_updated -- callbacks receive 2 ints a parameters (local_height,
server_height).  Usually they are equal but local_height may be larger
than server_height if the server we are connected to is lagging behind
other servers in terms of the latest blockchain blocks it's seen.

wallet_updated -- callbacks receive the 'wallet' instance as a
parameter. Client code can look at this instance parameter and ignore
the notification if the update was for a different wallet.

verified2 -- callbacks receive the 'wallet' instance as well as the other original 'verified' parameters of tx_hash, conf, height, etc..

The old 'updated' event will still be sent if callbacks have asked for it --
but the code will issue deprecated warnings using print_error each time
this happens.  Same goes for 'verified'

Old plugins will continue to function as will legacy code, but with
warnings.

This change will allow client code to better be informed about the
state changes of the wallet vs/ blockchain.

FWIW -- Electrum has done this also long ago.

Also changed:

The 'status' event now also sends a string  as a parameter indicating the status (eg 'disconnected' or 'connecting' or 'connected'). (previously it sent no parameters).

I grep'd through all the code and fixed everything to work with the new API, with the exception of android which eludes me.  Android will continue to build and function correctly it just will issue warnings so that the next developer maintaining it can upgrade to the new API. :)

Note that Android still is using the old 'updated' scheme and thus it
needs to be migrated over else it will issue deprecated warnings.

* small nit

* Deprecated the not very good 'verified' network callback

.. in favor of the more useful 'verified2' callback.

The old 'verified' callback is still supported but will issue warnings
if in verbose mode to stderr using print_error

* small nit