[Qt] If running from source, give the app an icon
- On macOS the app would get the generic "Python" icon (which was ugly).
Now, it detects if running from source and it picks up the icon '.icns'
file on darwin

- On Linux and Windows, it uses the icons/electron.ico file for the
icon.

This only takes effect in "running from source" mode and if the file is
present.

The other modes (built mode or srcdist) get their icons from the window
manager and work as before.