[Qt] On delete payment request, reset address label
If the address has no tx's and its label was the payment request's
label, it makes sense to also delete the label we gave it when we saved
the payment request.

Also found a bug in labels plugin where if user set a 'None' label it
would raise an exception.  Fixed.