[Qt] Fix #101 in CashShuffle branch Fix #968 in regular branch
Make the Receive tab ensure the address shown has no tx's and is new.

This check is done if the user has the 'use change address' flag set
in preferences which indicates they care about not re-using addresses.

If it's set, then the receive tab will always refresh itself with a
brand new address on the "new_transaction" network event iff a transaction
is received and the current address displayed happens to now have an
address history as a result of the new tx.

In addition, I modified the UI for the receive tab to be more intuitive
(I feel) and to have fewer glitches.

- The actively selected item in the request list always represents
what's in the form and vice-versa.

- The 'new' button has been renamed to "Clear" and is always
active/enabled

- The 'save' button is now only ever enabled if there are actual changes
to save.  Previously its behavior/enabled state was complex and erratic
and buggy.

- You can edit a request label in the tree widget directly and it ACTUALLY
has an effect and re-saves the request with the new label (previously it would
let you edit it in the tree widget but then the form above would not update properly
and the actual request would not get updated in the wallet).