[Qt] Remove call to main_window.deleteLater() on window close
This call is a bit superfluous and has the potential to cause bugs if
some plugin holds a reference to the main_window past its lifetime.

In the normal case, the C++ object does get deleted when the Python object is
finalized, which happens rather soon after window close.

If the python object does not get finalized this is a bug in whatever
code is holding the main_window object alive past the window close, and
needs to be tracked down.

However, having the C++ side raise an exception in the latter case was not really
good, so we removed this deleteLater() call as all it has the potential to
do is cause problems.