[Qt][Lib] Fix #1199 - Sometimes when turning on FX history, no update
The issue was the GUI wasn't being updated with (history_list.update)
-- this has been addressed.

In addition, the UI was wonky in that it would reset your exchange to the 0
position in the prefs exchange combo box because the exchagne comboBox.clear()
call, plus the addItems call would spam currentIndexChanged() events while we
were busy modifying the combobox, overriding preferences before we got a
chance to set the combo box back to its previous state.

Also, we now set the default exchange to "CoinGecko" since it supports
both history and regular and it also isn't deprecating the free history
rates API.

We also set the default currency to USD since we are US-centric jerks.
:)