[Qt] Cope better with bad or corrupt wallets from gui_last_wallet
Fixes #1189.

Previously the app would just present an error screen and abort if
config gui_last_wallet was a bad/corrupt or otherwise unknown wallet type.
This is not desired behavior as there was no reasonable way for a novice
user to resolve this situation.

Instead, in this situation we present them with the InstallWizard where
they get the opportunity to either select a different wallet file or
create a new one from scratch, as would be the expected UX.

This issue hit users "dual-booting" between EC SLP and regular but is
not limited to them -- any wallet corruption would result in an
unrecoverable state.