[Qt] Nit: Make sure QCompleter object of send payto has proper parent
Small nit: from the Qt docs the QCompleter needs a parent. In practice
python gc will keep it around until send tab deletion (window close) --
this is because it's referenced by a member in class payto_e.

But having top level qobjects dangling around without parents is
in general poor Qt-foo. So I made the QCompleter belong to the payto_e
to be more in-line with traditional C++/Qt.