[Lib] 2x speed boost on signing tx's if using Old_KeyStore wallets
These wallets are slow as-is for signing tx's with lots of inputs due to
the key stretching they use.

However, to add insult to injury, the key was being stretched twice while
signing.

So we made the `check_seed' function also return the stretched key in
case the caller needs it.  This speeds up tx signing on these admitedly
deprecated wallets by a factor of 2.