[Trezor] Fixups to bugs
- Bumped lib in requirements to 0.11.1
- Addeed transport string to places in UI where user must pick a device
to differentiate identical devices that have multiple transports
- Fixed some of the crash bugs related to "call in progress" -- there's
nothing we can do if the trezor lib is blocked in a call but we can at
least tell the user to unplug the trezor and try again.
- Fixed some of the inconsistent/bad messages in the install wizard when
restoring from seed.
- Some of the restore from seed flows are smoother now but still the
device can end up in a bad state if user hits cancel from the GUI due to
limitations of trezorlib