[Ledger] Fix the support for Ledger HW1/Nano legacy hardware w/o touching Nano S/Blue (#959)
* Fix the support for Ledger HW1/Nano legacy hardware

Ledger HW1/Nano have support for Bitcoin Cash in version 1.0.4, so check for
that. The call to getWalletPublicKey with cashAddr=True causes the legacy
devices to return an error we can use to detect it. The warning about cashddr
is suppressed as they have no display anyway. The 2FA dialog has been fixed by
converting the address to a legacy address.

* Add special case handling for Ledger HW.1/Nano legacy devices

* Ledger: Remove is_legacy and use of BTChipException for identification

* Ledger: Use finalizeInput instead of finalizeInputFull to support multiple outputs

* Ledger: Improve 2FA dialog and support testnet

Legacy Ledger codes are always generated from Legacy mainnet addresses. If
that does't match the current display address, also show the display
address.