[Qt][Lib] Follow-up to #1251 -- Undo assumption of always rich text
base_wizard may be some day used by a future platform that is not Qt.
Repect that lib/ makes no assumptions about UI capabilities.  So we
resort back to plain text and not rich text.

Also redid some of the wording/link and layout for the "failed to find a
device" install wizard message.