Windows: Make console work properly
This patch adds console functionality for the Windows build. As the Windows .exe
is a Windows subsystem executable it doesn't have a console. We solve this by
checking if there is a console of a parent process that we can attach to.
Additionally the --forceconsole parameter can be passed to force allocating a
new console if no parent console is found.