[Qt] Have 'Max' button in Send tab give visual cue when in 'max' mode
This is taken pretty much verbatim from Electrum commit 9b4e490e3cffe07c0f5a5e373b0834c6915baac1

The button has been made "checkable". It is "checked" when the send tab
has been set to "max" mode.  Once checked, it cannot be unchecked easily
(as before).  However, now the user at least has visual feedback that
he's entered "max" mode.

Note as always the send tab behavior is a bit non-obvious:

- Once you are in "max" mode the only way to leave it is to start to
manually enter an amount, at which point the max button becomes grayed out
(disabled) and "max" mode is left.  This disabling of the max facility
is a bit of a UX faux pas but we will address it in a later commit.

- If the "amount" edit is blank then you can toggle max mode again.

The above is the way it's always worked.  Just highlighting here how
non-obvious it is.  The only thing this commit does is give users a
visual cue that they are in "max" mode by having the button enter the
"checked" state when in "max" mode (on macOS this looks like a blue
highlighted button, on other platforms it looks like a pressed-down
button).