Console: Only show completions when tab is pressed without modifier keys
Ctrl+Tab / Ctrl+Shift+Tab can be used to navigate the main window tabs,
but as soon as the console is shown you can not switch to another tab. This
is because the console consumes all tab key events, even those with
modifiers. This patch makes the console ignore tab presses with modifier
keys.