Reworded FUD message from BIP39 seed dialog to not be FUD
Words like "compromised" and "we provide no guarantees" are FUD and
untrue. If anything the actual state of affairs is our seed format may
one day stop being supported becaused BIP39, like it or not, is pretty
much the standard format.

So, I rewrote the text to be factual and not a propaganda piece from a
bygone era.