AppImage: Copy libusb binary into image
pkg2appimage excludes libusb-1.0.so by default for no good reason:

https://github.com/AppImage/pkg2appimage/blob/83483c2971fcaa1cb0c1253acd6c731ef8404381/excludelist#L112

This can cause an issue when the AppImage loads the systems libusb but the
systems libusb in turn loads libudev from the AppImage. The kernel ABI for
libusb will not be changing so it is safe to bundle it into the AppImage.