Windows: Add --build to configure calls to prevent cross compilation misdetection
When building on Linux, we have to explicitly specify the build triplet because
otherwise cross compilation detection might break.

When the docker host has wine installed and binfmt_misc set up to execute .exe
files, cross compilation check inside the container will fail. That check just
compiles a .exe file and tests if it can execute it. If it can execute it, even
through wine / binfmt_misc, it assumes it is not cross compiling. Specifying
--build triplet will always skip that check and always assume cross compilation.