Windows: Fix zbar build on OSX hosts
We patch Makefile.am to enable static linking, but this doesn't get picked up on
OSX docker hosts. Turns out we patch that file _after_ calling autoreconf. When
building on a Linux host the Makefile correctly detects that Makefile.in needs
to be regenerated and does just that, but on an OSX host it fails to do that.
The solution for this is to just patch Makefile.am before calling autoreconf so
we directly generate the correct Makefile.in.