Windows: Recreate the ico file to support high resolutions
The previous icon had an unused alpha channel as well as 2 formats
(64 and 128) that Windows will never use for icon files, resulting in a
much larger file size than necessary. There was also some gradient
effect and some weird darkening on the edges.

The new icon uses only 24 bit color with sizes 256, 48, 32, 24 and 16.
Care was taken to have sharp color transitions on pixel boundaries
instead of a gradient between them.