Icons: Use recreated clock icons by the Simple Ledger devs
The clock icons were the last icons still stuck with low resolution.
We wanted to keep the original look instead of going with a flat
look like Electrum but the original vector files seem to have been
lost. We decided to recreate the icon so that it closely matches the
original when rendered at a low resoltion while giving more detail
when rendered at a higher resolution.

The Qt SVG engine seems to have an issue with these SVGs, rendering
them somewhat too dark. Until we can figure out a fix for this
problem we rasterize the icons.