Qt: Open URL in last wallet window that the user interacted with
The open url handler now tries to open the URL in the lastest window
that is still open that the user interacted with.

This means the ElectrumGui object now tracks which wallet window
the user last interacted with via the QApplication focusChanged signal.

That window is kept in a weak ref and if it's still alive, it is sent
the URL's data for pay_to_URI.

If that window has since died, windows[0] is the fallback.