CoinTex: Added a popup label to tell the user what to do next
On successful cointext PR retrieval and processing, the user still needs
to hit Preview or Send -- yet many users said the fact that they need to
do an additional step here was not obvious to them.  So we added a popup
label to tell them that they should review the tx and pointing to the
send button.