Lib: Schnorr slight refactor
Made the wallet functions: mktx, make_unsigned_transaction, and cpfp not
require the schnorr_flag to be passed (it can be passed to override
config, though). Instead, the wallet just uses whatever the config
setting is for that wallet.  This makes older plugins automatically have
schnorr capability if they use these functions for schnorr.

- Also enforced sign_schnorr=False in CashShuffle to explicitly suppress
schnorr in case API ever changes.