Receive tab: Added optional OP_RETURN parameter
This allows one to construct URLs and/or QR codes in the receive tab
with either 'op_return' or 'op_return_raw' in the URL query string.

This parameter is supported by Electron Cash's send tab already, so it's
useful to be able to construct OP_RETURN URLs if sending the receive
request to Electron Cash.

Other wallets will just ignore this parameter if they don't support the
'op_return' or 'op_return_raw' query string arguments.