QR Scanner: Added informative message for Linux users having trouble
This message is displayed for Linux users if the QR scanner fails due to
Qt unable to detect a camera.

This situation arises on Linux only because PyQt5 <5.12 forgot to package
a critical Qt library: libQt5MultimediaGstTools.so.

PyQt5 versions >= 5.12.2 have already addressed this problem, but on
earlier versions the library is missing and thus Linux users with
perfectly usable cameras may get an error when attempting to use the QR
scanner.

This commit adds an informative message to the generic error, telling
users how to install the latest PyQt5 in order to access their camera.