Tx Dialog: Warn user if attempting to broadcast low fee tx
This is to prevent pathological uses of the app where user ends up in tx
dialog and broadcasts a tx that the bitcoind on the other end accepts,
but the rest of the network does not, leaving his funds in limbo (at
least while connected to that ElectrumX/bitcoind combo).

Can happen on air gapped systems with mixed Schnorr/non-Schnorr installs
guessing fee incorrectly, leading to low fee on broadcast.

The warning is shown in the tx dialog before broadcasting.  Presumably
this is to prevent the user from broadcasting a tx that may never
confirm, so they can go back to square 1 (send tab of the network-facing
system) and start again.

Future proof added: The warning is not shown if it was already shown in the
send tab on sign (uses the tx's 'ephemeral' dict).