Tx Dialog: Added context menu for inputs and outputs!
Closes #1439
Partially addresses #1229

This basically brings the tx dialog closer to blockchain explorer
functionality (but not quite) -- all prevouts in the inputs text edit
are right-clickable and you can do things like open the tx (in another
tx dialog), copy the txid:n, copy the address, view the address history
(if address in wallet), etc.

This also applies to the outputs textedit.

The new functionality is a power feature because it requires right-click
-- if somebody wants I can make it more "front and center" in the UI by
changing the inputs/outputs text edits in the TxDialog into
QTreeWidgets...