Tx Dialog: All wallet addresses are now clickable
This partially addresses my dream to make Electron Cash comparable to a
block explorer.

So addresses in the tx dialog that belong to the wallet (eg change or
receiving) are now underlined and are linkified (their foreground color
is not changed, however).

Clicking on them leads to the Address detail window (same window you get if
you right-click on an address in the address book and click 'Details').

Closes #1229