Fix #1223 "Disappearing Menu Actions" bug (#1456)
* cleaned up imports for cashshuffle/qt.py
* Fixed the menu bug..

It was due to signals/slots for objects living in different threads
being connected to lambdas... and I think PyQt ended up firing the slots
in the wrong threads.  As well as the `disconnects` being in the wrong 
threads maybe led to the wrong signal/slots being disconnected. Either way 
it was out of spec to send signals/slots across threads to lambdas as the receiver 
was the wrong thread.