Fix #1460: HW Wallet 'keysstore.thread' objects never exit
There was a bug where the hw wallet's 'TaskThread' objects (used to talk
to the device, presumably, in an asynchronous fashion) would never exit
and would remain running, in a blocked state, as zombie threads forever.

They are now ended on wallet close in a synchronous fashion.

This situation was the cause of bug #1460, where the wallet would be
closed and its C++ objects would be cleaned up but the thread was still
running, doing work.

By ending the thread in a synchronous fashion, the hope is this bug is
prevented.