Address tab: Added right-click 'Copy Legacy Address'
If app is in cashaddr mode, shows a second item 'Copy Legacy Address'
after the primary 'Copy Address'.  If in Legacy mode, the second item
becomes 'Copy Cash Address'.

Also refactored the copy-to-clipboard code to be a generic instance
function of class ElectrumWindow. It pops up a tooltip 'Text copied to
clipboard' after copying to clipboard.  This was to reduce the amount of
boilerplate code.