TxDialog: Added 'Mined in block' field to the dialog
This label also has a link, which if you click it, it brings you to the
block explorer page for the tx in question.

This might be useful to some users, as a quick way to see the tx on the
blockchain. Otherwise user has to copy txhash manually to get to block
explorer page for tx.

This "Mined in block" label only appears in the tx dialog for confirmed
tx's we know the heights of (tx's in the wallet)