Build: Protect against PGP keyserver poisoning attack
This patch removes the downloading of the Python release signing
keys from keyservers. Instead we just import these from files in
the repository. This speeds up the build and makes it more reliable.

The source of the release signing keys is:

https://www.python.org/downloads/#pubkeys

These keys might change in the future and the file has to be updated
then, but that should only happen when we update our Python version.

See this link for details on the keyserver attack:

https://gist.github.com/rjhansen/67ab921ffb4084c865b3618d6955275f