Fonts: Improve color emoji support on Windows and Linux
We switch Windows to use FreeType to be able to render the subset of
Noto Color Emoji we are using. This also enables us to use the same
Schnorr sigil on Linux as on Windows.

We name the font 'Segoe UI Emoji' because that will override the emoji
font on Windows. Sadly Qt has no way to override font fallback so this
is the only way to do this other than binary patching or recompiling Qt.

The font now lives under gui/qt/data/emojis.ttf and gets installed by
setup.py.