UTXOList: Added default sort order of Amount, Descending
 Also the UTXOList now 'remembers' the last sort order the user
selected on a per-wallet basis (but if never specified, defaults to the
somewhat more useful Amount, Descending).

The mechanism for remembering the sort order for a MyTreeWidget is in
the base class, so it can be re-used in other tabs. For now, UTXIList
(coins) is the only tab that will use this mechanism.

Other small nits are also contained in this commit.