Cash Accounts Implementation + Many Bug Fixes (#1505)
Closes #1035 .

The main UI components to it are in:

- **Contacts tab**
- Addresses tab, right click -> Register new.. (to register to a specific address)
- Tools -> Lookup Cash Account
- Receive tab (to register a cash account on the fly to the receive address there in the tab)

What it has:

- *Full verification* -- users are never presented with any results unless they are first SPV verified. They are also warned when a cash account in the contacts tab is no longer verified (due to reorg?) and won't offer it as an option for pay-to.

- Cash Account Registration Auto-Detection™ - Detects cash accounts related to the wallet (that is, registration tx's for the wallet) whenever it "sees" a tx. This can either be a tx that came through the normal wallet "add_transaction" path or if the user explicitly opens a tx, etc.

- The lookup facility is pretty robust and resilient to all sorts of shenanigans.

What it lacks:

- 100% Send tab niceties: Right now the only way to send to a cash account from the send tab is to **first** add it as a contact in the contacts tab.


This last Send Tab bit is a somewhat complex UI flow but I will do it.  I just wanted to get this PR merged in.

As I was working on this I fixed/found/cleaned up a ton of bugs and made the code tighter overall.

This PR includes all of that as well, so I would like to merge it fairly soon and move forward from master with work on Cash Accounts send tab perfection.