Qt: Performance fix to History tab for wallets with many txs when verifying
There is some performance bug in Qt 5.11 & 5.12 whereby updating the
history list (eg when a tx becomes verified and we need to update its
icon and status in the history) was horrendously slow for a history list
with many thousands of tx's in it.

See the EatBCH wallet here:

38ty1qB68gHsiyZ8k3RPeCJ1wYQPrUCPPr

This commit adds a workaround to make the history list "update_item"
call fast again by first removing the item from the QTreeWidget before
adding it again after update.  This turns out to be much faster than
updating it in place.

Each update_item call took 500ms (!!) for this EatBCH wallet but is now
sub-ms in terms of cost.

User UX for removing, updating, then adding the item again is exactly the same
except not it's fast and the GUI doesn't hang... so win/win.