Cash Accounts: Added full payto suppot in send tab
This works well and doesn't clobber OpenAlias.

The code basically checks any cashaccounts specified in the payto text
edit and if detected, makes sure they are verified. If they are, great,
it then proceeds to fill in their emoji and address.

If verification "fails" or drops out while the send tab is up (say due
to an unlikely reorg), the <address> is deleted.

This allows user to free-form specify any cash account in the send tab.

Multiline is supported too eg:

someguy#1000, 0.122
someotherguy#123, 1.23
John#456.22; , 1.45
qqppltassweqeqwqweqdasq, 3.45

So you can mix and match bitcoin cash addresses and cash accounts and
nothing bad happens.. it all "just works", Etc...

Note again nothing is every auto filled-in and addresses are deleted if
the cashaccount is not verified or if verification subsequently
"disappears".  So it should be really safe to use.