CashAcconts: small optimization to get_minimal_chash
On 'minimal_ch_cache' miss, we proceed to also check the "processed
block cache" when computing the minimal chash for a cash account,
and return the computed value immediately from the cached processed
block immediately.  This is a performance speedup (saving reduntant
network lookups when we have the info already) as well as a UX
improvement.