CashAcct: Show [Pend] pending registrations in contacts tab
Thanks to imaginary_username for this great suggestion. We show pending
registrations as nam#???.??? in the contacts tab with a gears icon next
to it. This is so that right after a registration is issued, the user
gets an immediate sense that their registration is "going through". The
pseudo-contact displayed is ephemeral and completely there only as a UI
nicety. It doesn't get saved to disk and cannot be deleted, edited, or
"payed-to".  It simply exists as an advisory UI element.