InstallWizard: Make the "Back" button when prompting password Cancel
The back button did nothing since the stack was effectively empty and/or
causing the same dialog to reappear.

In this case, rather than rewrite the wizard to suppot "back" from the
password entry screen, we instead made the "Back" button be a "Cancel"
button. This closes the Install Wizard immediately.