Qt: Disable shortcuts, clear menubar on window close
This is because on macOS sometimes the menubar actions (and their
shortcuts) stick around for a while after the window is closed but
before Python GC kicks in.  Due to the shared menubar on macOS, bad
things can happen if user hits a shortcut on a closed window.  The dead
window comes back to life but is empty.. and other nonsense.

By disabling all shortcuts on window close and deleting the window's
menubar object, this can no longer happen.

The bug this fixes is rare and esoteric but I've seen it once or twice
today when testing other things.