Qt: Send tab, added a small little '?' help button to the Pay to
The tool button is small and unobtrusive but provides help for users
about how to use the (rather complex) payto edit.

The help text displayed is the same as if one were to click on the "Help
Label" to the left of the payto (but these help labels rarely are
noticed by users, whereas it is hoped the help button will be more
obvious).