Fonts: Use a more safe fontconfig override
There were issues with font fallback at least on Arch Cinnamon and
on a Debian 9 I've been using in Qubes. The new fonts.xml should be
more safe, but also the user might need to install the Bitstream Vera
font to work optimally. This is because we blacklist the DejaVu font
because it includes bad emojis and the Bitstream Vera font is
essentially DejaVu without the emojis.