CashAccounts: Added parsing & integration of cashacct: URIs (#1547)
* CashAccounts: Added parsing of cashacct: URIs

This involved the minimal possible changes:

- electron-cash script accepts cashacct: URIs
- PayToEdit knows how to deal with them
- Main window pay_to_URI modified
- web.parse_URI can parse them
- web.create_URI can create them

* Added desktop integration for cashacct: URL type

- macOS Info.plist
- Windows setup.exe .nsi installer
- Linux .desktop file (AppImage, etc)

* Added optional %-decode for macOS, others

Sometimes the '#' character comes in as %-encoded if the URL was opened
from the desktop.  For cashacct: URIs we should tolerate this and
%-decode the cashacct name.

* Windows: Remove cashacct URI from registry on uninstall

I forgot to delete this key. Oops! :)

* Updated linux xdg desktop.xml

Added the bitcoincash: and cashacct: schemes, and refreshed some of its
data to be more current.