Addresses tab (Qt): Disallow Right-click -> Request payment on used addr
Closes #1552. Previous to this commit -- the UI offered "request
payment" on used addresses which was an inconsistent UI action.  The
receive tab would pop up with a request on a DIFFERENT address than the
one we right-clicked on.  This is because the receive tab as of Electron
Cash 4.0 basically outright refuses to generate duplicate payment
requests or payment requests on used addresses (unless you are in an
imported wallet where it has no choice).

So, rather than permit this UI inconsistency, we instead gray-out the
"request payment" option for used addresses, but still offer it for
unused addresses (where it makes sense).

See #1552.