CashAccounts: Lookup results "View tx..." link changed to "Details..."
It's probably more useful to be able to click on details for lookup
results (to get the big detail window) for a particular cash account
than it is to see the registration tx.

Note that from the details window it's possible to see the registration
tx ultimately.  So I feel this is more complete.