Fix #1559: Crash due to packaging errors of currencies.json
If currencies.json mentions currency exchange providers not present in
exchange_rate.py (which apparently happened recently on iOS), the app
will crash with a KeyError.

Since this is a potential foot-gun, we guard against it with catching
KeyError exceptions when attempting to access a class that does not
exist in exchange_rate.py and safely ignoring the error.

Closes #1559