Always define MyTreeWidget filter_columns class attribute
It turns out that hitting Ctrl+F when some plugins were loaded led to an
uncaught exception because they neglected to define `filter_columns` as
a class attribute.

We provide a default definition of the empty list (which implies no
filter) for the base class.

The base class was accessing this attribute in the filter() method so it
really did need this attribute always defined.

This fixes crash bugs on extant "in-the-wild" copies of the scheduled_payments
plugin.

In this commit we also guard against plugins doing unorthodox things such as
not definining the .filter() method for their searchable_list tabs
(Inter Wallet Transfer plugin suffered from this API misuse).