iOS: Nit in setup code to pass the correct argv to UIApplicationMain
It doesn't make much practical difference right now -- but we were
passing an incomplete argv (we omitted the executable name in argv[0]).
Not entirely sure if this matters, but this has been now fixed for
completeness (who knows if some underyling cocoa objc lib expects arv[0]
to be the actual executable name and otherwise is unhappy).