Bugfix in i18n: Fixed bug from an earlier commit in detecting system language
Commit cba803acc78e9a7686c44eef183830cbbfa717d8 clobbered the module
`locale` with an import of `lib/locale`, inadvertently. This broke the
i18n.get_system_language_match function which relied on the Python
module `locale` to have attribute getdefaultlocale().

This fixes the bug.

Interestingly this error was silently ignored and new installs would
just default to English... which is something some users wanted anyway.
:)