OSX: Fixups for Mojave visual & font glitches (#1580)
It turns out the way we've been building Electron Cash on El Capitan using PyQt 5.11 led to visual glitches on Mojave.  Some widgets didn't render quite right and font weights were too bold, among other minor problems.

Qt 5.12 is the first version of Qt to actually look good on Mojave.

But there is a problem with Qt 5.12 for compatibility: it no longer supports El Capitan.

As a result we do the following:

- Upgraded `make_osx` to use the same requirements file as the other platforms (which uses PyQt 5.12.3); this fixes the Mojave glitch problem
- *HOWEVER* 8-12% of macOS users are still on El Capitan and Qt 5.12 won't run on El Capitan, so we also support building on El Capitan systems for El Capitan. This will produce a `.dmg` with a `-elcapitan-macosx.dmg` suffix.
- As such, `make_osx` can auto-detect if it is being build on El Capitan and it will use an alternate `requirements-binaries.txt` file which uses PyQt 5.11.
- So, in summary: For a few releases we should release both the regular (built on Sierra, works on Mojave, uses Qt 5.12) `.dmg` as well as the `-elcapitan-macosx.dmg` (built on El Capitan, uses Qt 5.11, works flawlessly on everything up to Mojave).

Additionally, this commit provides a workaroud and a warning should the user run the `-elcapitan-macosx.dmg` on a Mojave system (but it still allows them to proceed).

This commit also slightly refactors `electroncash_gui/qt/__init__.py` to put some of the pre-App-setup initialization stuff into a separate method.

---

* MacOS: Mojave workaround for PyQt 5.11

We warn the user that PyQt 5.11 + Mojave is a bad time. We also try to
apply some hacky workarounds to mitigate some of the font rendering
issues with Qt 5.11 on Mojave.

* OSX: Fix to use alternate requirements-binaries for El Capitan builds

We use an alternate, PyQt 5.11 requirements-binaries for El Capitan
builds.

All other newer MacOS will use PyQt5.12.3.

* OSX: Nits to Mojave Qt5.11 fixup code

* make_osx: Added more pre-requisites for brew

* tweak to make_osx