Qt TxDialog: Allow copy/paste of LockTime, layout nit, and fix #1483
- LockTime label was not selectable by mouse. Maybe user wants to
copy/paste the locktime. Fixed.

- LockTime label had funny formatting after it by hardcoding a \n
(newline) in the label. We address this formatting more properly by
using an appropriate setContentsMargin on the QHBoxLayout that immediately
follows it, rather than the unorthodox approach of relying on a newline
within a QLabel for spacing.

- Fix issue #1483: The "Broadcast" button remains enabled after a
successful broadcast for a time until the wallet "sees" the broadcasted
tx. In some situations the button stays enabled for a few seconds
tempting the user to press it again.  We implement a cool-down period
after a success status is returned from the ElectrumX server on
broadcast whereby the button is forced disabled for a minimum of 5 seconds
after a broadcast success.  If we hear from the server later, it stays
disabled permanently.  If we don't, the button comes back to life
allowing the user to try and broadcast again.