Network: Fix a race condition, also fix tx input fetcher bug #1583
There was a race condition in network.py's 'message_id' incrementation
logic. It used self.message_id += 1 which is not guaranteed thread-safe
in Python.  So instead we use a util.Monotonic object to increment the
message_id which is thread-safe if constructed with locking=True.

Also, there was a subtle bug in the tx input fetcher which would
sometimes raise an uncaught exception if there are no interfaces
available when fetching inputs (closes #1583).