Build READMEs: Delete outdated/defunct readmes
We no longer support manual building outside of the `build.sh` script as
supporting that mechanism is too burdensome.

If people wish to build manually by issuing 16 different unix commands,
they are free to read the `_build.sh` script and do things manually that
way.

(What's ore: It's not terribly useful to build outside of the official `build.sh`
script.)