Bumped version to 4.0.10 for dev, added RELEASE-NOTES
We bumped the version to 4.0.10 because a recent commit added stuff to
the API and we want plugin authors to be able to differentiate versions.
4.0.10 is still in development, but we might as well bump the version
now.

We modified authors to be "The Electron Cash Developers" since having
just one name on there doesn't do justice to all the other contributors.