Ensure Electron Cash codebase is not run with -O (optimize)
The Electron Cash lib can fail subtly and annoyingly if run with
PYTHONOPTIMIZE (-O).  A lot of the logic in the codebase depends on
'assert' not being a no-op.

The benefit to using -O in Python is *extremely* limited, since, well,
Python is pretty slow anyway and compile-time or run-time optimization
only goes so far. However, for this codebase, the negatives to using -O
-- namely 'assert' becoming a no-op -- leads to incorrect execution.

So, we test for -O and enforce it in bitcoin.py (one of the files that
heavily depends on correct assert behavior).

This is a pet peeve since it always bothered me that some client code
could inadvertently enable -O and thus cause incorrect execution in
Electron Cash.