Add SLP Awareness to prevent token burn (#1602)
Closes #1494, see PR for details.

 Commit log:
---

* initial refactor
* more refactor
* nits
* nit
* tweaks
* added tests
* fixed tests to pass
* nits in slp test
* added OP_RETURN build ScriptOutput tests
* nit

* Added slp.WalletData stub. Needs rethinking

* tweaks

* Tweak

* added WalletData rough sketch

* Moved the module lib/slp.py into a package in lib/slp/

This will allow us to add various pieces such as the
slp_validator_0x1.py and other things in the future.

* renamed lib/slp/base.py -> lib/slp/slp.py

* Added miminal slp.WalletData implementation

This implementation does the bare minimum of maintainting fast data
structures and keeps track of the things we need right now for SLP
awareness to prevent tokens.

We give this provisional data layout slp_data_version '3.9' which should
cause EC-SLP to rebuild slp data when switching between this and EC-SLP.

In the future we will add more structure to this data when we enable
validation.

Regardless -- we need to do away with the EC-SLP data layout as it is
somewhat inefficient.

* got rid of an extraneous TODO

* Hooked SLP WalletData class into wallet.py

wallet.slp is the attribute. It seems to work well and keep track of SLP
utxos. Next up -- GUI support.

* Added SLP awareness to the UI

- wallet.get_utxos() defaults to excluding SLP tokens from all spends.
This should make it impossible to burn SLP tokens from a UI action.

- Indicate SLP tokens on coins in the coins tab using green.

- Indicate SLP tokens in CashShuffle by labeling them as such in column
5 (they never get shuffled, even if they have sufficient amount on them)

* nit

* nit

* Reoarganized how the data is saved to wallet.storage

Everything goes into a great big dict named 'slp' in wallet.storage.
This way someone editing the file can have an easier time.

* nit

* fixed typo

* Minor nit and speedup

* slp.WalletData.load: Simplified to make the method more readable

We flattened it out and rely on asserts more since now the main EC
backend guarantees 'assert' is available.

The code in .load is more readable now, IMHO