Network Dialog: Workaround Qt bug on some Linux
On some Linux systems, the Network Dialog in the Qt UI was showing up
"empty" as a window with no contents and transparent. Issuing a
C++ QDialog::update() (QDialog.update in Python) call to the instance
on showEvent causes a repaint and fixes the issue.

Not sure what the cause of the issue was and not sure why only on Linux
-- but whatever works.