iOS: Add SLP-awareness to the UI
Coins tab now shows unspendable SLP tokens as highlighted in green (as
if they were frozen coins) but rather than saying "Frozen" it says "SLP
Token".

Previous to this commit iOS didn't show the SLP tokens at all but the
balance did reflect them... so this was necessary before release of the
new iOS build to the App Store.