i18n: Add support for pgettext and npgettext with msgctxt
Some words need additional context to be correctly translated, like
in the case of the current "None" translation string. For example
in German you need to know the gender of what is referenced to
correctly translate it. This is different in different parts of the
UI and we previously used the wrong form in some places.

Context will allow us to have multiple tranlsations of the same
words depending on where they are used.

In gettext, pgettext and similar functions would be used for this.
Sadly Python only added these in version 3.8. To support them with
lower Python versions, we take advantage of those versions not
being aware that the separator of msgctxt and msgid is \x04 and
instead look up a msgid that is combined with the msgctxt.