Addresses tab: hide all empty change addrs, not just used ones
Rationale: there is very little need for users to care about the future
empty addresses that will be automatically chosen for them. For the
"Received" list it can be useful to right click on an empty to grab an
address, set a label, etc; not so for the "Change" list.

(If they really feel the need, however, the empty unused addrs are just
one more click away.)