iOS fix #1651 - app can crash if prefs screen up when wallet closed
Wallet gets auto-closed after 2 minutes of app being in background (as
per iOS rules).  When user resumes app, the prefs screen is still up and
was querying wallet properties for schnorr (but self.wallet was None).

Since we never found a good solution for this we need to guard all
property queries on the wallet with: `if self.wallet` boilerplate.

This was normally done but I forgot to do this for the new schnorr prefs
screen.

It has now been addressed. This fixes issue #1651.