Qt history tab: Added workaround for persitent 'Unknown' balances
This can happen if the input's script is somewhat complex and the
pruned_txo_values cache has to try and download prevouts and figure out
the addresses in question.

The fact that pruned_txo_values has stuff in it causes the
get_wallet_delta function to return "None", which makes the GUI render
"Unknown" in the balance column.

This situation would eventually correct itself after the wallet settles.

The code we added was to schedule a GUI refresh sometime after the last
tx was verified and after the wallet settled, to clearn the "Unknown"
balances from the history tab.

This is a follow-up to PR #1620.

These persistent 'Unknown' balances are caused by the code added in that
PR, and this workaround is now needed as a result.