iOS: PR #1670 broke iOS a little bit, this fixes it.
iOS version needed to be updated to support the new Daemon() constructor
arguments having a 'plugins' arg now (iOS passes None here as no plugins
ship with iOS).  The daemon is ok with this.. at least the way in which
iOS uses the daemon class it is.

Anyway.. the new API was making the iOS app not work anymore, this fixes
that.