Improvements to Export History
This closes #1679

Exporting history has been improved:

- It now returns a fee (if known) -- best efforts are made to examine
prevout_tx (and if the inputs are from wallet & known), it uses those
values to calculate fee.  This calculation is done in a thread with a
250ms timeout (because it does involve deserializing tx's, etc),
leveraging the pre-existing  fee calc code in transaction.py.  It
doesn't go out to the network to do this, though.

- Fees that are so-computed while exporting history are saved back to
the wallet in self.tx_fees

- The export history CLI has been updated as well as the Qt GUI to
always export fee (if known)

Also in this commit:

- Some small nits/fixups to Transaction class