Follow-up to 54ba44463239a413ebd8755f60e9d7fcaae68761
Defend against reorg: Since we don't take the wallet.lock for
deadlock reasons (we rely on a thread that needs the lock), we have to
defend against the wallet losing its history while we are exporting!
(Can happen in pathological but rare circumstances if a reorg happens or
other funny business while the export_history function is running).

This commit detects that situation and deals with it appropriately.