Network dialog: Fix not showing .onions before switching to the Proxy tab
The state of the "Use tor" checkbox is used to determine whether to show
.onion servers in the server list. This used to not be saved and only
initialized when switching to the proxy tab, causing .onion server to not
be shown.

This patch changes it so that the state of the checkbox is saved. It also
removes the proxy_settings_changed code that resets the box to unchecked
when the dialog opens. When the user chose to use tor, we disable the
ability to change the settings instead.