Build: Fix regression causing newer glibc to be pulled into AppImage
In fd3701fd6c64a726fd201e53e77c0cf4e33a5598 we added a new git version
to the AppImage build image to support shallow submodule updates.

The problem with this is that it causes a newer glibc to be pulled
into the image which breaks older Linux distributions.

As a solution to this we use a git PPA by the Ubuntu Git maintainers
that is compiled for Ubuntu 16.04:

https://launchpad.net/~git-core/+archive/ubuntu/ppa?field.series_filter=xenial

Due to a change in git 2.22 shallow clones will fail when using the
default git protocol version, so we set protocol.version to version
2 instead. See:

https://public-inbox.org/git/20191013064314.GA28018@sigill.intra.peff.net/