Nit in gui/qt/util.py: `all` doesn't need a wrapping list
This is in MyTreeWidget.filer.  The `all()` function doesn't need a list
inside to work.  Can just do a free-form generator and it does what you
expect.  Not sure if it Python optimizes this away anyway but.. it
bothered my sensibilities that CPython might create a temporary list
when a real generator would work there.