Fix #1789: Disallow user delete last address in Imported wallet
Imported privkey and Imported address wallets allow users to delete the
last address in the UI, leading to a wallet with 0 addresses.  The rest
of the app does not cope well with this situation (can generate random
exceptions).

This commit disallows such an action from the Qt UI.

This closes #1789.