Fix #1774: Tolerate spurious 'error' messages from server.
Fulcrum sends a spurious 'error' message to the client right before
terminating the connection due to idle timeout.  Normally this can't
happen if the client is behaving, however in some circumstances, such as
on mobile where it may go dormant, or on desktop systems that are
thrashing due to low memory and never get to send 'server.ping', this spurious
message was leading to an exception being thrown in Electron Cash.

This corner case happens exceptionally rarely, but it will be addressed
in Fulcrum soon: See https://github.com/cculianu/Fulcrum/issues/20

In the meantime, we made the network 'interface' class more
resilient/tolerant of the situation.

Also, we added additional guards against malformed server messages in
interface.py.

This closes #1774, and fixes the related #1794.