Fix #1738 - Handle protobuf serialization errors for payment requests
This may happen in rare cases when the payment request contains
out-of-range values such as an amount that doesn't fit into a 64-bit
integer in C++.

We need to catch these exceptions and handle them properly, which this
commit addresses.

Where it makes sense we display the error to the user (usually a
ValueError).

Also in this commit -- we added an optional argument, on_exc to the
web.parse_URI function which is in case of payment request exceptions
in the subthread.  It is called in the context of the main thread
using the new do_in_main_thread mechanism.

Closes #1738.