OSX: Added entitplements.plist file.
This file is to be used for manually re-signing the built app using
hardened runtime. See: https://github.com/pyinstaller/pyinstaller/issues/4629

We will add logic to the build script to automate this process at a
later time -- but that logic will need this file anyway.  For now we
codesign the build binary again using hardened runtime and specifying
this plist file.