OSX: Fixups to osx build script to support hardened runtime for Catalina
MacOS Catalina requires codesigning with the "hardened runtime" option.
This requires the use of the 'entitlements.plist' file, among other
things.

We add support to sign in this way. Requires Mojave+ host, otherwise the
codesigning will not be hardened (and the app will not be notarized
properly).

Also various other small fixups to the make_osx script.