Tor: Make process shutdown more robust
This first instructs the Tor process to shut down via the control
port, then waits for the process to exit. This will handle almost
all cases.

If the process does not shut down within 1 second we send it a
shutdown signal via the OS and wait indefinitely for the process
to exit.

This also correctly handles ProcessLookupError, which can occur
when killing a process.