Event: Replace forced main thread invocation with function decorator
The Event class currently forces every method call to happen on the
main thread. There are some use cases where this is not a good idea,
so this should be removed. Instead we add the function decorator
@in_main_thread that invokes the decorated function in the main
thread. Every event handler that touches GUI objects or needs to
run in the main thread for some other reason needs to be decorated.