Tor: Make the messages refer to a specific Tor binary
The messages now refer either an integrated Tor, a system Tor or
just Tor. When no Tor binary is detected, Tor will be deactivated
and there will be a message about this in the tooltip of the settings.

This also changes the Fusion "Click here to resolve" to "Click
this popup to resolve" because "here" does not make it clear whether
to click on the popup or on the Fusion button it is pointing to.