Prompt to start Tor only if user has enabled autofuse for wallet
This fixes issue #1970.  It turns out we shouldn't be prompting users if
they want to enable Tor unconditionally, even if autofuse is disabled.
Many users have no idea what CashFusion is and may find this prompt
annoying.  We only kick off the task to maybe prompt the user if no Tor
port is found only after the user has actually enabled autofuse (or if
autofuse is on due to wallet settings).

The net result is:

- On startup, if wallet is set to autofuse in config, the prompt appears
as before

- On manual enable of wallet from the "no autofuse" state, the prompt
appears (this behavior is new).

- If wallet has autofuse disabled, no prompt appears for that wallet
if Tor is missing (this behavior fixes #1970).

Closes #1970.