Make signature verification more compatible with other libs
Electron Cash normally signs a message by prepending a "preamble" to the
message of the form: b"\x18Bitcoin Signed Message:\n" + length + message

Not all libs do this, in particular golang bchec does not add any
preamble and just signs the original message hash.

This commit allows Electron Cash to try both methods when verifying a
signed message.  It first tries the Bitcoin way (with preamble), and if
that fails, it tries the non-preamble.  If either succeed, the message
is considered verified for the given address.

This allows Electron Cash to verify messages generated by other
libraries.  Previous to this commit, a valid signed message generated by
bchec in golang would fail in Electron Cash.

Security note:

The fact that we try to verify against two possible message hashes does
not decrease the security of this software in any appreciable way -- and
instead just increases the UX for users asked to verify messages
generated by other software.